requirements.txt
flask==2.0.1
flask-cors==3.0.10
psycopg2-binary==2.9.1
python-dotenv==0.19.0
gunicorn==20.1.0
pydantic==1.10.13
sqlalchemy==1.4.41
PyJWT==2.8.0
Werkzeug==2.0.3 